MSEBufferController to own file, BufferController abstract class to own file, EventHandler abstraction integrated with BufferController. TODO: Other event handlers should also implement this.
